requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.4
pydantic==2.5.3
pydantic-settings==2.1.0
sentence-transformers==2.2.2
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy<2.0.0
pdfplumber
pdfminer.six>=20200726
python-docx
